Analyze Power Sensitivity to VDD and Frequency

Performs a sensitivity analysis by calculating dynamic power over a range of VDD and frequency values, by fixing switching activity and capacitance.

analyze_power_sensitivity(alpha_fixed, C_fixed, V_DD_range, f_range)

Arguments

alpha_fixed

The fixed average switching activity factor (e.g., average alpha).

C_fixed

The fixed total capacitance (in Farads, F).

V_DD_range

A numeric vector of supply voltage values (in Volts, V) to test.

f_range

A numeric vector of clock frequency values (in Hertz, Hz) to test.

Value

A matrix where rows correspond to VDD values and columns to frequency values. Each element contains the estimated dynamic power (in Watts, W).

Examples

VDDs <- seq(0.8, 1.2, by = 0.1) # 0.8V to 1.2V
Freqs <- seq(0.5e9, 2e9, length.out = 4) # 0.5 GHz to 2 GHz
power_matrix <- analyze_power_sensitivity(0.15, 60e-12, VDDs, Freqs)
print(power_matrix)
#>           f_0.5 GHz  f_1 GHz f_1.5 GHz f_2 GHz
#> VDD_0.8 V 0.0014400 0.002880 0.0043200 0.00576
#> VDD_0.9 V 0.0018225 0.003645 0.0054675 0.00729
#> VDD_1 V   0.0022500 0.004500 0.0067500 0.00900
#> VDD_1.1 V 0.0027225 0.005445 0.0081675 0.01089
#> VDD_1.2 V 0.0032400 0.006480 0.0097200 0.01296