Apply Zero Transition Coding (Bus-Inversion Coding)

Optimizes dynamic power by minimizing the number of bit transitions (flips) between two consecutive data words. It transmits the inverted word if the resulting activity is lower, adding an inversion bit indicator.

apply_ztc_coding(data_vector_prev, data_vector_curr)

Arguments

data_vector_prev

A numeric vector (0s and 1s) representing the previously transmitted word.

data_vector_curr

A numeric vector (0s and 1s) representing the current word to be transmitted.

Value

A list containing:

optimized_data: The vector transmitted (either the original or the inverted word).

inversion_bit: The bit indicator (1 if inversion was performed, 0 otherwise).

transitions_saved: The number of transitions saved by the optimization.

Examples

D_prev <- c(0, 0, 0, 0, 0, 0, 0, 0)
D_curr_high_activity <- c(1, 1, 1, 1, 1, 1, 1, 1) # 8 transitions
# Inverted D_curr is 00000000 (0 transitions)
result <- apply_ztc_coding(D_prev, D_curr_high_activity)
print(result)
#> $optimized_data
#> [1] 0 0 0 0 0 0 0 0
#> 
#> $inversion_bit
#> [1] 1
#> 
#> $transitions_saved
#> [1] 8
#>